mm: centralize kernel phys/virt address translation in mm module
Introduce `kernel/src/mm.rs` with a single `KERNEL_VIRT_OFFSET` constant
(sourced from `CONFIG_KERNEL_VIRT_OFFSET` kconfig) and two const fns
`kernel_phys_to_virt` / `kernel_virt_to_phys`.

Remove the duplicated, arch-gated versions that were scattered across
`allocator/mod.rs`, `allocator/buddy/heap.rs`, `devices/virtio.rs`, and
`arch/aarch64/mmu.rs`. All call sites now import from `crate::mm`.

Also move `CONFIG_KERNEL_VIRT_START` from `arch/arm/arm64/Kconfig` to the
top-level `arch/Kconfig` as `CONFIG_KERNEL_VIRT_OFFSET`, making the offset
available to all architectures instead of being aarch64-only.

+++ kernel/src/boards/qemu_virt64_aarch64/init.rs
@@ -90,7 +90,7 @@ crate::define_peripheral! {
      )),
 }
 
-pub const DRAM_BASE: u64 = mmu::kernel_phys_to_virt(0x4000_0000);
+pub const DRAM_BASE: usize = mmu::kernel_phys_to_virt(0x4000_0000);
 
 crate::define_pin_states!(None);
